+++ upcloud_api/cloud_manager/__init__.py
@@ -4,6 +4,7 @@ from upcloud_api.api import API
 from upcloud_api.cloud_manager.firewall_mixin import FirewallManager
 from upcloud_api.cloud_manager.host_mixin import HostManager
 from upcloud_api.cloud_manager.ip_address_mixin import IPManager
+from upcloud_api.cloud_manager.lb_mixin import LoadBalancerManager
 from upcloud_api.cloud_manager.network_mixin import NetworkManager
 from upcloud_api.cloud_manager.object_storage_mixin import ObjectStorageManager
 from upcloud_api.cloud_manager.server_mixin import ServerManager
@@ -12,14 +13,15 @@ from upcloud_api.cloud_manager.tag_mixin import TagManager
 
 
 class CloudManager(
-    ServerManager,
-    IPManager,
-    StorageManager,
     FirewallManager,
-    TagManager,
-    NetworkManager,
     HostManager,
+    IPManager,
+    LoadBalancerManager,
+    NetworkManager,
     ObjectStorageManager,
+    ServerManager,
+    StorageManager,
+    TagManager,
 ):
     """
     CloudManager contains the core functionality of the upcloud API library.